legal_assistant/requirements.txt
Flask==2.1.0
Werkzeug==2.0.3
google-generativeai==0.3.1
python-dotenv==0.19.2
PyPDF2==3.0.1
pdfminer.six==20221105
python-docx==0.8.11
Pillow==9.3.0
pytesseract==0.3.10
pdf2image==1.16.3
